chore: moved render params to separate method
